I did not end up using docker

+++ FDM/FDM.c
@@ -5,7 +5,7 @@
 
 
 //Time Steppper for the Runge Kutta Method using an Explicit Integration Scheme
-matrix *step_RK_E_h(matrix *y0, matrix **Y_h, float t0, float h, Interp_function Intrpl, ODE_function odefcn_h, matrix *a, matrix *b, matrix *c, flexBody *body){
+matrix *step_RK_E_h(matrix *y0, matrix **Y_h, double t0, double h, Interp_function Intrpl, ODE_function odefcn_h, matrix *a, matrix *b, matrix *c, flexBody *body){
 
     matrix **k = (matrix **)malloc(sizeof(matrix *) * b->numCols);
 
@@ -21,7 +21,7 @@ matrix *step_RK_E_h(matrix *y0, matrix **Y_h, float t0, float h, Interp_function
     for(int i = 0; i < b->numCols; i++){
         y_intrpl = Intrpl(Y_h, t0 + c->data[0][i] * h);
 
-        for(int j = 0; j< i, j++;){
+        for(int j = 0; j < i, j++;){
             matrix_add(temp, matrix_scalar_mul(k[j], a->data[i][j]));
         }
         //y_h = (y_intrpl = Intrpl(Y_h, c(i));)
